+++ tests/test_vedalang_cli.py
@@ -473,11 +473,8 @@ class TestViz:
             {
                 "id": "toy_states_alt",
                 "veda_book_name": "TOYSTATESALT",
-
-                "base_year": 2025,
-
+                "year_set": "pathway_2025_2035",
                 "currency_year": 2024,
-
                 "region_partition": "toy_states",
             }
         )